# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bolão Copa 2026 - World Cup Betting Pool API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  inviteToken: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  participant: "participant",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type UserSummaryRole =
  (typeof UserSummaryRole)[keyof typeof UserSummaryRole];

export const UserSummaryRole = {
  admin: "admin",
  participant: "participant",
} as const;

export interface UserSummary {
  id: number;
  email: string;
  name: string;
  role: UserSummaryRole;
  totalPoints: number;
  betsSubmitted: number;
}

export interface Group {
  id: number;
  name: string;
  description?: string;
}

export interface Team {
  id: number;
  name: string;
  flag?: string;
  fifaCode: string;
  groupId?: number;
  group?: Group;
}

export type GroupWithTeams = Group & {
  teams: Team[];
};

export interface CreateGroupRequest {
  name: string;
  description?: string;
}

export interface CreateTeamRequest {
  name: string;
  flag?: string;
  fifaCode: string;
  groupId?: number;
}

export interface CreateInviteRequest {
  email: string;
  expiresInDays?: number;
}

export interface Invite {
  id: number;
  email: string;
  token: string;
  used: boolean;
  expiresAt: string;
  createdAt: string;
}

export interface InviteResponse {
  invite: Invite;
  inviteUrl: string;
}

export interface InviteValidationResponse {
  valid: boolean;
  email: string;
}

export type CreateMatchRequestStage =
  (typeof CreateMatchRequestStage)[keyof typeof CreateMatchRequestStage];

export const CreateMatchRequestStage = {
  group: "group",
  round_of_32: "round_of_32",
  round_of_16: "round_of_16",
  quarterfinal: "quarterfinal",
  semifinal: "semifinal",
  third_place: "third_place",
  final: "final",
} as const;

export interface CreateMatchRequest {
  homeTeamId?: number;
  awayTeamId?: number;
  homePlaceholder?: string;
  awayPlaceholder?: string;
  groupId?: number;
  matchDate: string;
  stage: CreateMatchRequestStage;
  matchNumber?: number;
  venue?: string;
}

export type UpdateMatchResultRequestStatus =
  (typeof UpdateMatchResultRequestStatus)[keyof typeof UpdateMatchResultRequestStatus];

export const UpdateMatchResultRequestStatus = {
  upcoming: "upcoming",
  live: "live",
  finished: "finished",
} as const;

export interface UpdateMatchResultRequest {
  /** @minimum 0 */
  homeScore: number;
  /** @minimum 0 */
  awayScore: number;
  status: UpdateMatchResultRequestStatus;
  homeTeamId?: number;
  awayTeamId?: number;
}

export type MatchStage = (typeof MatchStage)[keyof typeof MatchStage];

export const MatchStage = {
  group: "group",
  round_of_32: "round_of_32",
  round_of_16: "round_of_16",
  quarterfinal: "quarterfinal",
  semifinal: "semifinal",
  third_place: "third_place",
  final: "final",
} as const;

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus];

export const MatchStatus = {
  upcoming: "upcoming",
  live: "live",
  finished: "finished",
} as const;

export interface Match {
  id: number;
  homeTeamId?: number;
  awayTeamId?: number;
  homeTeam?: Team;
  awayTeam?: Team;
  homePlaceholder?: string;
  awayPlaceholder?: string;
  groupId?: number;
  group?: Group;
  matchDate: string;
  stage: MatchStage;
  matchNumber?: number;
  venue?: string;
  homeScore?: number;
  awayScore?: number;
  status: MatchStatus;
  bettingDeadline: string;
  createdAt: string;
}

export interface Bet {
  id: number;
  userId: number;
  matchId: number;
  homeScore: number;
  awayScore: number;
  points?: number;
  exactScore?: boolean;
  correctOutcome?: boolean;
  correctGoalDiff?: boolean;
  createdAt: string;
  updatedAt: string;
}

export type MatchWithBet = Match & {
  userBet?: Bet;
};

export interface SubmitBetRequest {
  matchId: number;
  /** @minimum 0 */
  homeScore: number;
  /** @minimum 0 */
  awayScore: number;
}

export type BetWithMatch = Bet & {
  match?: Match;
};

export type BetWithUserUser = {
  id: number;
  name: string;
};

export type BetWithUser = Bet & {
  user?: BetWithUserUser;
};

export interface RankingEntry {
  rank: number;
  userId: number;
  userName: string;
  totalPoints: number;
  exactScores: number;
  correctOutcomes: number;
  bonusPoints: number;
  betsSubmitted: number;
  totalMatches: number;
}

export type UserRankingDetail = RankingEntry & {
  recentBets?: BetWithMatch[];
};

export interface SubmitBonusBetsRequest {
  champion: string;
  topScorer?: string;
}

export interface BonusBet {
  id: number;
  userId: number;
  champion: string;
  topScorer?: string;
  championPoints?: number;
  topScorerPoints?: number;
  locked: boolean;
  createdAt: string;
  updatedAt: string;
}

export type GetTeamsParams = {
  groupId?: number;
};

export type GetMatchesParams = {
  stage?: GetMatchesStage;
  status?: GetMatchesStatus;
  groupId?: number;
};

export type GetMatchesStage =
  (typeof GetMatchesStage)[keyof typeof GetMatchesStage];

export const GetMatchesStage = {
  group: "group",
  round_of_32: "round_of_32",
  round_of_16: "round_of_16",
  quarterfinal: "quarterfinal",
  semifinal: "semifinal",
  third_place: "third_place",
  final: "final",
} as const;

export type GetMatchesStatus =
  (typeof GetMatchesStatus)[keyof typeof GetMatchesStatus];

export const GetMatchesStatus = {
  upcoming: "upcoming",
  live: "live",
  finished: "finished",
} as const;