# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Bolão Copa 2026 - World Cup Betting Pool API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: getTeams
      tags: [teams]
      summary: Get all teams
      parameters:
        - name: groupId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
  /matches:
    get:
      operationId: getMatches
      tags: [matches]
      parameters:
        - name: stage
          in: query
          schema:
            type: string
            enum: [group, round_of_32, round_of_16, quarterfinal, semifinal, third_place, final]
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, live, finished]
        - name: groupId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchWithBet"
  /matches/{matchId}:
    get:
      operationId: getMatch
      tags: [matches]
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchWithBet"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bets:
    post:
      operationId: submitBet
      tags: [bets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitBetRequest"
      responses:
        "200":
          description: Bet submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bet"
        "400":
          description: Deadline passed or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bets/match/{matchId}:
    get:
      operationId: getMatchBets
      tags: [bets]
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bets for match (visible after deadline)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BetWithUser"
...
  /admin/invite:
    post:
      operationId: createInvite
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResponse"
...
  /admin/groups:
    post:
      operationId: createGroup
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /admin/teams:
    post:
      operationId: createTeam
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /admin/matches:
    post:
      operationId: createMatch
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchRequest"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /admin/matches/{matchId}/result:
    put:
      operationId: updateMatchResult
      tags: [admin]
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMatchResultRequest"
      responses:
        "200":
          description: Result updated, scores recalculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Match not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
      required: [id, name]
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        flag:
          type: string
        fifaCode:
          type: string
        groupId:
          type: integer
        group:
          $ref: "#/components/schemas/Group"
      required: [id, name, fifaCode]
    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
        flag:
          type: string
        fifaCode:
          type: string
        groupId:
          type: integer
      required: [name, fifaCode]
    CreateInviteRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        expiresInDays:
          type: integer
          default: 7
      required: [email]
    InviteResponse:
      type: object
      properties:
        invite:
          $ref: "#/components/schemas/Invite"
        inviteUrl:
          type: string
      required: [invite, inviteUrl]
    Invite:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        token:
          type: string
        used:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, email, token, used, expiresAt, createdAt]
...
    CreateMatchRequest:
      type: object
      properties:
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        homePlaceholder:
          type: string
        awayPlaceholder:
          type: string
        groupId:
          type: integer
        matchDate:
          type: string
          format: date-time
        stage:
          type: string
          enum: [group, round_of_32, round_of_16, quarterfinal, semifinal, third_place, final]
        matchNumber:
          type: integer
        venue:
          type: string
      required: [matchDate, stage]
    UpdateMatchResultRequest:
      type: object
      properties:
        homeScore:
          type: integer
          minimum: 0
        awayScore:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [upcoming, live, finished]
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
      required: [homeScore, awayScore, status]
    Match:
      type: object
      properties:
        id:
          type: integer
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
        homePlaceholder:
          type: string
        awayPlaceholder:
          type: string
        groupId:
          type: integer
        group:
          $ref: "#/components/schemas/Group"
        matchDate:
          type: string
          format: date-time
        stage:
          type: string
          enum: [group, round_of_32, round_of_16, quarterfinal, semifinal, third_place, final]
        matchNumber:
          type: integer
        venue:
          type: string
        homeScore:
          type: integer
        awayScore:
          type: integer
        status:
          type: string
          enum: [upcoming, live, finished]
        bettingDeadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, matchDate, stage, status, bettingDeadline, createdAt]
    MatchWithBet:
      allOf:
        - $ref: "#/components/schemas/Match"
        - type: object
          properties:
            userBet:
              $ref: "#/components/schemas/Bet"
    SubmitBetRequest:
      type: object
      properties:
        matchId:
          type: integer
        homeScore:
          type: integer
          minimum: 0
        awayScore:
          type: integer
          minimum: 0
      required: [matchId, homeScore, awayScore]
    Bet:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        matchId:
          type: integer
        homeScore:
          type: integer
        awayScore:
          type: integer
        points:
          type: integer
        exactScore:
          type: boolean
        correctOutcome:
          type: boolean
        correctGoalDiff:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, matchId, homeScore, awayScore, createdAt, updatedAt]
...
    BetWithUser:
      allOf:
        - $ref: "#/components/schemas/Bet"
        - type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
              required: [id, name]
...