# lib/api-zod/src/generated/types/betWithUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bolão Copa 2026 - World Cup Betting Pool API
 * OpenAPI spec version: 0.2.0
 */
import type { Bet } from "./bet";
import type { BetWithUserUser } from "./betWithUserUser";

export type BetWithUser = Bet & {
  user?: BetWithUserUser;
};