# lib/api-zod/src/generated/types/createInviteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bolão Copa 2026 - World Cup Betting Pool API
 * OpenAPI spec version: 0.2.0
 */

export interface CreateInviteRequest {
  email: string;
  expiresInDays?: number;
}