# lib/api-zod/src/generated/types/invite.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bolão Copa 2026 - World Cup Betting Pool API
 * OpenAPI spec version: 0.2.0
 */

export interface Invite {
  id: number;
  email: string;
  token: string;
  used: boolean;
  expiresAt: Date;
  createdAt: Date;
}